+++ problem4j-spring-web/src/main/java/io/github/problem4j/spring/web/resolver/HttpRequestMethodNotSupportedProblemResolver.java
@@ -46,7 +46,7 @@ public class HttpRequestMethodNotSupportedProblemResolver extends AbstractProble
    * @since 1.2.0
    */
   public HttpRequestMethodNotSupportedProblemResolver() {
-    this(ProblemFormat.identity());
+    super(HttpRequestMethodNotSupportedException.class);
   }
 
   /**
@@ -55,7 +55,13 @@ public class HttpRequestMethodNotSupportedProblemResolver extends AbstractProble
    *
    * @param problemFormat the problem format to use
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link
+   *     io.github.problem4j.spring.web.config.ProblemBeanPostProcessor ProblemBeanPostProcessor}
+   *     now assigns the {@link ProblemFormat} after construction; use {@link
+   *     #HttpRequestMethodNotSupportedProblemResolver()}
    */
+  @SuppressWarnings("removal")
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public HttpRequestMethodNotSupportedProblemResolver(ProblemFormat problemFormat) {
     super(HttpRequestMethodNotSupportedException.class, problemFormat);
   }