+++ problem4j-spring-web/src/main/java/io/github/problem4j/spring/web/resolver/MissingServletRequestPartProblemResolver.java
@@ -54,7 +54,7 @@ public class MissingServletRequestPartProblemResolver extends AbstractProblemRes
    * @since 1.2.0
    */
   public MissingServletRequestPartProblemResolver() {
-    this(ProblemFormat.identity());
+    super(MissingServletRequestPartException.class);
   }
 
   /**
@@ -63,7 +63,13 @@ public class MissingServletRequestPartProblemResolver extends AbstractProblemRes
    *
    * @param problemFormat the problem format to use
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link
+   *     io.github.problem4j.spring.web.config.ProblemBeanPostProcessor ProblemBeanPostProcessor}
+   *     now assigns the {@link ProblemFormat} after construction; use {@link
+   *     #MissingServletRequestPartProblemResolver()}
    */
+  @SuppressWarnings("removal")
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public MissingServletRequestPartProblemResolver(ProblemFormat problemFormat) {
     super(MissingServletRequestPartException.class, problemFormat);
   }