+++ problem4j-spring-web/src/main/java/io/github/problem4j/spring/web/resolver/ResponseStatusProblemResolver.java
@@ -44,7 +44,7 @@ public class ResponseStatusProblemResolver extends AbstractProblemResolver {
    * @since 1.2.0
    */
   public ResponseStatusProblemResolver() {
-    this(ProblemFormat.identity());
+    super(ResponseStatusException.class);
   }
 
   /**
@@ -52,7 +52,13 @@ public class ResponseStatusProblemResolver extends AbstractProblemResolver {
    *
    * @param problemFormat the problem format to use
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link
+   *     io.github.problem4j.spring.web.config.ProblemBeanPostProcessor ProblemBeanPostProcessor}
+   *     now assigns the {@link ProblemFormat} after construction; use {@link
+   *     #ResponseStatusProblemResolver()}
    */
+  @SuppressWarnings("removal")
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public ResponseStatusProblemResolver(ProblemFormat problemFormat) {
     super(ResponseStatusException.class, problemFormat);
   }