+++ problem4j-spring-webmvc/src/main/java/io/github/problem4j/spring/webmvc/resolver/NoHandlerFoundProblemResolver.java
@@ -45,7 +45,7 @@ public class NoHandlerFoundProblemResolver extends AbstractProblemResolver {
    * @since 1.2.0
    */
   public NoHandlerFoundProblemResolver() {
-    this(ProblemFormat.identity());
+    super(NoHandlerFoundException.class);
   }
 
   /**
@@ -53,7 +53,13 @@ public class NoHandlerFoundProblemResolver extends AbstractProblemResolver {
    *
    * @param problemFormat the problem format to use
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link
+   *     io.github.problem4j.spring.web.config.ProblemBeanPostProcessor ProblemBeanPostProcessor}
+   *     now assigns the {@link ProblemFormat} after construction; use {@link
+   *     #NoHandlerFoundProblemResolver()}
    */
+  @SuppressWarnings("removal")
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public NoHandlerFoundProblemResolver(ProblemFormat problemFormat) {
     super(NoHandlerFoundException.class, problemFormat);
   }